refactor: split internal/actions/split.go into logical files
